+++ pingtop.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 
 import click
+import urwid
 import threading
 from concurrent.futures import ThreadPoolExecutor
 from ping import do_one
 import time
+from gui import screen, tablebox, get_palette, old_signal_keys
 
 
 hosts = {"baidu.com": {}, "alipay.com": {}}
@@ -24,24 +26,35 @@ def forever_ping(dest, index_flag):
         time.sleep(1)
 
 
+def global_input(key):
+    if key in ("q", "Q"):
+        # TODO Ctrl-C
+        event.clear()
+        raise urwid.ExitMainLoop()
+    else:
+        return False
+
+
+mainloop = urwid.MainLoop(
+    urwid.Frame(urwid.Filler(tablebox)),
+    palette=get_palette(),
+    screen=screen,
+    unhandled_input=global_input,
+)
+
+
 @click.command()
 def multi_ping():
     global hosts
+    pool = ThreadPoolExecutor(max_workers=len(hosts))
+    event.set()
+    for index, host in zip(range(len(hosts)), hosts):
+        last_future = pool.submit(forever_ping, host, index)
+
     try:
-        pool = ThreadPoolExecutor(max_workers=len(hosts))
-        event.set()
-        for index, host in zip(range(len(hosts)), hosts):
-            last_future = pool.submit(forever_ping, host, index)
-        last_future.result()
-    except KeyboardInterrupt:
-        event.clear()
-        print("shutdown!")
-        for host in hosts:
-            print(host)
-            print(host, min(hosts[host]["rtts"]))
+        mainloop.run()
     finally:
-        pool.shutdown(wait=False)
-        print("shuted!")
+        screen.tty_signal_keys(*old_signal_keys)
 
 
 if __name__ == "__main__":
